feat: implement feed management and login flow hooks
- Added `useFeedManagementPage` hook for managing feeds and folders, including CRUD operations and state management.
- Introduced `useLoginFlow` hook to handle the login process with server URL validation and credential submission.
- Refactored timeline-related hooks by creating `useTimelineSelection` for managing article selection state.
- Updated `useTimeline` to integrate new selection and auto-mark read functionalities.
- Created utility functions for validating login server URLs and credentials in `login.ts`.
- Added session management functions in `session.ts` for encoding credentials and converting session data.
- Implemented article preview utilities in `articlePreview.ts` for normalizing article data and reconciling folder/feed names.
- Removed deprecated sync module and adjusted related tests to reflect changes in API structure.
