--- src/lib/sync/index.ts
@@ -1,2 +1,0 @@
-export { fetchUnreadItemsForSync } from '@/lib/api/itemsSync';
-export { reconcileTimelineCache } from '@/lib/storage/timelineCache';
